fix file lock error

+++ libultimate/api.py
@@ -8,6 +8,7 @@ class API():
     def __init__(self, ryujinx_path: str):
         self.game_state_path = os.path.join(ryujinx_path, 'sdcard/libultimate/game_state.json')
         self.command_path = os.path.join(ryujinx_path, 'sdcard/libultimate/command.json')
+        self.command_ok_path = os.path.join(ryujinx_path, 'sdcard/libultimate/command.ok.json')
 
     def read_state(self):
         with open(self.game_state_path, 'r') as f:
@@ -17,6 +18,12 @@ class API():
             return game_state
 
     def send_command(self, command):
-        with open(self.command_path, 'w') as f:
-            json.dump(command, f)
+        if not os.path.isfile(self.command_ok_path):
+            with open(self.command_path, 'w') as f:
+                json.dump(command, f)
+            # create ok file
+            with open(self.command_ok_path, 'w') as f:
+                pass
+        else:
+            print("Warning: command not sent, previous command not acknowledged")
